Convert directives from comment early
pipe_convert_directives_from_comment() was done at the beginning of
pipe_gen_problems(), but it's better to do it early in a separate pipe
segment so that pipe_annotate() can check for the presence of tags and
refs.

pipe_convert_markup_to_directives() now sets directives directly instead
of appending comments.

And pipe_annotate() warns if a tag or ref already exists.