The dynamic site structure, used for the gogamespace.com WordPress site, is now much simpler.
There is now only one page that is used for displaying a collection of
problems. It is a real WordPress page, not just a static HTML file.

The URL of a collection is now

    /training-module-collection/?collection=by_filter/foo

or

    /training-module-collection/?collection=by_id/abcd1234

I've added a hook to GeneratePress' functions.php that takes the query
parameter and loads the relevant title strings and problem JSON data
from the JavaScript file that corresponds to that collection.

That way we don't need to use jQuery or D3 to load the JSON, and I think
that it should be faster as well because the browser can cache the JS
files.