generate problems: more fine-grained debugging
When writing debug information to a generated problem's GC[], it was
only possible to write all the metadata or none. Now the user can
specify whether or not to write tags, refs and the tree location (file +
index).

The reason for this change is that I have files with long paths that
squashed the Go board. But these files are for books, so they also have
refs; in that case it's sufficient to display the refs.