Debugging enhancements
When generating problems, no longer offer the option to write metadata
to the GC[]. It already exists in the SGJ; the WGo viewer can use that
to display debugging data.

Therefore the Debug and GameInfo plugins were deleted.

gogame-munge-refs has been expanded and renamed to gogame-munge-sgj. It
can also skip "secret" problems completely and make the problem
filenames relative to avoid clutter.