site improvements
GoGameTools::Porcelain::SiteWrite->write_collection_file() now just
writes a JSON file, not a JavaScript file.

The collection is read using the Fetch API. This makes the code a lot
simpler, both on the server and on the client side.

Moved code from the collection index page (outside this repo) to
initTraining() in training.js. So it's more centralized and the index
page is simpler too.