refactored the pattern annotation code
Refactored pipe_annotate() into GoGameTools::Porcelain::Annotate.

It now annotates the source trees directly with #tagname:p; these are called
"pattern tags". sctipt/gogame-set-pattern-tags first traverses the collection
and removes all pattern tags, then annotates the collection anew.

Previously the annotations were added while generating problems. This had the
disadvantage that they could go out of sync with the source problem trees. You
had to always regnerate the annotations if you changed the source trees, which
was time-consuming.

So now the annotation is not done in script/gogame-gen-problems or
script/gogame-gen-site but in a completely separate step while you edit the
source trees.

script/gogame-set-pattern-tags also makes a {{ note }} of annotations that were
skipped because the node was not marked with TE[]. So when editing the source
trees you can search for those notes and possibly choose to add TE[] to that
node.