refactored making linear stories
There JSON file where the story is specified can now contain different
types of instructions.

The "path" instruction is how it worked up to now. You specify a "from"
and a "to" node and this node list is added to the story.

The new "node" instruction just adds a single node to the story.

The new "traverse" instruction goes to the "from" node and does a
depth-first traversal, adding a node list for each variation.